+++ src/main/java/com/coolerpromc/archerythings/network/packet/ServerBoundQuiverMenuPacket.java
@@ -1,6 +1,7 @@
 package com.coolerpromc.archerythings.network.packet;
 
 import com.coolerpromc.archerythings.ArcheryThings;
+import com.coolerpromc.archerythings.compat.AccessoriesHelper;
 import com.coolerpromc.archerythings.component.ModDataComponents;
 import com.coolerpromc.archerythings.item.ModItems;
 import com.coolerpromc.archerythings.screen.quiver.QuiverMenu;
@@ -11,6 +12,7 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.SimpleMenuProvider;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.neoforged.fml.ModList;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
 public record ServerBoundQuiverMenuPacket() implements CustomPacketPayload {
@@ -21,7 +23,7 @@ public record ServerBoundQuiverMenuPacket() implements CustomPacketPayload {
     public static void handle(ServerBoundQuiverMenuPacket packet, IPayloadContext context){
         context.enqueueWork(() -> {
             if (context.player() instanceof ServerPlayer player){
-                if (player.getItemBySlot(EquipmentSlot.CHEST).is(ModItems.QUIVER.get()) || player.getItemBySlot(EquipmentSlot.CHEST).has(ModDataComponents.STORED_QUIVER.get()) || player.getItemBySlot(EquipmentSlot.LEGS).has(ModDataComponents.STORED_QUIVER.get())){
+                if (player.getItemBySlot(EquipmentSlot.CHEST).is(ModItems.QUIVER.get()) || player.getItemBySlot(EquipmentSlot.CHEST).has(ModDataComponents.STORED_QUIVER.get()) || player.getItemBySlot(EquipmentSlot.LEGS).has(ModDataComponents.STORED_QUIVER.get()) || ArcheryThings.isQuiverEquipped(player)){
                     player.openMenu(new SimpleMenuProvider((id, inv, playerEntity) -> new QuiverMenu(id, inv, player, null), Component.translatable("item.archerythings.quiver")), buf -> buf.writeBoolean(false));
                 }
             }